+++ CHANGELOG.md
@@ -11,12 +11,13 @@ key and app attestation**, pinning down the last unnecessarily moving parts:
 
 * Rename `AttestationValidator` -> `AttestationVerifier` to align with wording (and introduce typealias, but marked as deprecated)
 * Rename `verifyKeyAttestation` -> `verifyAttestation` (and introduce delegate, but marked as deprecated)
+    * `CertificateIssuer` now has `AttestationResult.Verified` as receiver for the new function.
+    * The deprecated function stays as it is.
 * Allow `CertificateIssuer` to throw instead of returning a KmmResult
 * Ship a default OID to identify the attestation proof.
 * Add defaults for keyConstraints and nonce validity duration.
 * Transmit device names inside CSR on a best-effort basis
 * Constrain challenge issuing wrt. validity duration: No more params can be specified, but informational adding of time zone is still allowed.
-* `CertificateIssuer` now has `AttestationResult.Verified` as receiver.
 
 **It also includes behavioural changes to the Android and iOS attestation defaults:**